urbvan/requirements/base.txt
Django==2.0.4
django-rest-framework==0.1.0
djangorestframework==3.8.2
marshmallow==2.15.0
factory-boy==2.10.0
